+++ Block/Adminhtml/System/Config/BoomerangVersion.php
@@ -4,10 +4,9 @@ declare(strict_types=1);
 namespace Basicrum\Analytics\Block\Adminhtml\System\Config;
 
 use Basicrum\Analytics\Model\Config;
-use Magento\Config\Block\System\Config\Form\Field;
 use Magento\Framework\Data\Form\Element\AbstractElement;
 
-class BoomerangVersion extends Field
+class BoomerangVersion extends ReadOnlyField
 {
     protected function _getElementHtml(AbstractElement $element): string
     {